Refactor authentication and category management
- Updated appsettings.json to nest Admin secret under Admin section.
- Enhanced RegisterDto and LoginDto with data annotations for validation.
- Modified CreateCategoryDto and UpdateCategoryDto to include validation attributes.
- Introduced OrderStatus enum for better order state management.
- Refactored AuthService to use IOptions for JWT configuration and added cancellation tokens to async methods.
- Updated CategoryService and OrderService to support cancellation tokens and improved error handling.
- Changed monetary fields in Order and OrderItem entities from double to decimal for precision.
- Added new exception classes: BusinessRuleException, ConflictException, and NotFoundException for better error handling.
- Updated repository interfaces and implementations to include cancellation tokens.
- Created a migration to alter database schema for monetary fields and nullable descriptions.

+++ EcomApi.Domain/Entities/OrderItem.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace EcomApi.Domain.Entities;
 
 public class OrderItem
@@ -8,7 +6,7 @@ public class OrderItem
     public int OrderId { get; set; }
     public int ProductId { get; set; }
     public int Quantity { get; set; }
-    public double UnitPrice { get; set; }
+    public decimal UnitPrice { get; set; }
 
     public Order? Order { get; set; }
     public Product? Product { get; set; }
